+++ CHANGELOG.md
@@ -5,11 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## 0.22.0
 
 ### Changed
 
-- Move Alarm plugin code to tomate-gtk repository
+- Move notify plugin to tomate-gtk repository
+
+## 0.21.0
+
+### Changed
+
+- Move alarm plugin to tomate-gtk repository
 
 ## 0.20.0